+++ src/Driver.h
@@ -292,10 +292,10 @@ public:
 
       msg2->toTuple(&res2, &numRedn2);
 
-      int numParticleCopies = *(int*)(res2[2].data);
-      int numParticleShares = *(int*)(res2[3].data);
-      int maxPESize = *(int*)(res2[0].data);
-      int sumPESize = *(int*)(res2[1].data);
+      long numParticleCopies = *(long*)(res2[2].data);
+      long numParticleShares = *(long*)(res2[3].data);
+      long maxPESize = *(long*)(res2[0].data);
+      long sumPESize = *(long*)(res2[1].data);
       float avgPESize = (float) universe.n_particles / (float) CkNumPes();
       float ratio = (float) maxPESize / avgPESize;
       bool complete_rebuild = (config.flush_period == 0) ?
@@ -303,7 +303,7 @@ public:
           (iter % config.flush_period == config.flush_period - 1);
 
       if (iter + 1 == config.num_iterations) complete_rebuild = false;
-      CkPrintf("[Meta] n_subtree = %d; timestep_size = %f; numPSParticleCopies = %d; numPSParticleShares = %d; sumPESize = %d; maxPESize = %d, avgPESize = %f; ratio = %f; maxVelocity = %f; rebuild = %s\n", n_subtrees, timestep_size, numParticleCopies, numParticleShares, sumPESize, maxPESize, avgPESize, ratio, max_velocity, (complete_rebuild? "yes" : "no"));
+      CkPrintf("[Meta] n_subtree = %d; timestep_size = %f; numPSParticleCopies = %ld; numPSParticleShares = %ld; sumPESize = %ld; maxPESize = %ld, avgPESize = %f; ratio = %f; maxVelocity = %f; rebuild = %s\n", n_subtrees, timestep_size, numParticleCopies, numParticleShares, sumPESize, maxPESize, avgPESize, ratio, max_velocity, (complete_rebuild? "yes" : "no"));
       //End Subtree reduction message parsing
 
       paratreet::postIterationFn(universe, proxy_pack, iter);
