+++ src/Subtree.h
@@ -31,7 +31,7 @@ public:
   std::vector<Node<Data>*> empty_leaves;
   Real load;
 
-  int n_total_particles;
+  long n_total_particles; // global count: 64-bit
   int n_subtrees;
   int n_partitions;
 
@@ -52,7 +52,7 @@ public:
 
   std::vector<Particle> flushed_particles; // For debugging
 
-  Subtree(const CkCallback&, int, int, int, TCHolder<Data>,
+  Subtree(const CkCallback&, long, int, int, TCHolder<Data>,
           CProxy_Resumer<Data>, CProxy_CacheManager<Data>, DPHolder<Data>, bool);
   Subtree(CkMigrateMessage * msg){
     delete msg;
@@ -124,7 +124,7 @@ public:
 };
 
 template <typename Data>
-Subtree<Data>::Subtree(const CkCallback& cb, int n_total_particles_,
+Subtree<Data>::Subtree(const CkCallback& cb, long n_total_particles_,
                        int n_subtrees_, int n_partitions_, TCHolder<Data> tc_holder,
                        CProxy_Resumer<Data> r_proxy_,
                        CProxy_CacheManager<Data> cm_proxy_, DPHolder<Data> dp_holder,
