+++ src/Writer.h
@@ -6,14 +6,14 @@
 #include <vector>
 
 struct Writer : public CBase_Writer {
-  Writer(std::string of, int n_particles);
+  Writer(std::string of, long n_particles);
   void receive(std::vector<Particle> ps, Real time, int iter);
   void write(CkCallback cb);
 
 private:
   std::vector<Particle> particles;
   std::string output_file;
-  int total_particles = 0;
+  long total_particles = 0;
   int cur_dim = 0;
   int iter_ = 0;
   Real time_ = 0;
@@ -23,7 +23,7 @@ private:
 struct TipsyWriter : public CBase_TipsyWriter {
   TipsyWriter(std::string of, BoundingBox b);
   void receive(std::vector<Particle> ps, Real time, int iter);
-  void write(int prefix_count, CkCallback cb);
+  void write(long prefix_count, CkCallback cb);
 
 private:
   std::vector<Particle> particles;
@@ -31,7 +31,7 @@ private:
   BoundingBox box;
   int iter_ = 0;
   Real time_ = 0;
-  void do_write(int prefix_count);
+  void do_write(long prefix_count);
 };
 
 #endif /* _WRITER_H_ */
